Simplify samples: TypeScript-only, use @microsoft/opentelemetry API
- Remove JavaScript samples and v1/ folder nesting
- Flatten structure to samples/src/*.ts
- Update all imports from @azure/monitor-opentelemetry to @microsoft/opentelemetry
- Replace useAzureMonitor/shutdownAzureMonitor with useMicrosoftOpenTelemetry/shutdownMicrosoftOpenTelemetry
- Restructure options to use azureMonitor.azureMonitorExporterOptions nesting
- Update README with configuration tables and clean up boilerplate
