+++ samples/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@microsoft/opentelemetry-samples",
+  "private": true,
+  "version": "1.0.0",
+  "description": "Microsoft OpenTelemetry distribution samples for TypeScript",
+  "engines": {
+    "node": ">=20.0.0"
+  },
+  "scripts": {
+    "build": "tsc",
+    "prebuild": "rimraf dist/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Azure/opentelemetry-distro-javascript"
+  },
+  "author": "Microsoft Corporation",
+  "license": "MIT",
+  "dependencies": {
+    "@microsoft/opentelemetry": "latest",
+    "dotenv": "latest",
+    "@opentelemetry/api": "^1.9.0",
+    "@opentelemetry/resources": "^2.2.0",
+    "@opentelemetry/semantic-conventions": "^1.38.0",
+    "@opentelemetry/sdk-trace-base": "^2.2.0",
+    "@opentelemetry/exporter-trace-otlp-http": "^0.208.0"
+  },
+  "devDependencies": {
+    "@types/node": "^20.0.0",
+    "rimraf": "latest",
+    "typescript": "~5.9.3"
+  }
+}
