+++ samples/src/cloudRole.ts
@@ -5,7 +5,7 @@
  * @summary Demonstrates how to set Cloud Role Name and Cloud Role Instance using OpenTelemetry Resource attributes.
  */
 
-import { useAzureMonitor } from "@azure/monitor-opentelemetry";
+import { useMicrosoftOpenTelemetry } from "@microsoft/opentelemetry";
 import { resourceFromAttributes } from "@opentelemetry/resources";
 import { ATTR_SERVICE_NAME } from "@opentelemetry/semantic-conventions";
 import "dotenv/config";
@@ -17,17 +17,17 @@ async function main(): Promise<void> {
     "service.instance.id": process.env.OTEL_SERVICE_INSTANCE_ID || "my-instance",
   });
 
-  const options = {
+  useMicrosoftOpenTelemetry({
     resource: customResource,
-    azureMonitorExporterOptions: {
-      connectionString:
-        process.env.APPLICATIONINSIGHTS_CONNECTION_STRING || "<your connection string>",
+    azureMonitor: {
+      azureMonitorExporterOptions: {
+        connectionString:
+          process.env.APPLICATIONINSIGHTS_CONNECTION_STRING || "<your connection string>",
+      },
     },
-  };
+  });
 
-  useAzureMonitor(options);
-
-  console.log("Azure Monitor configured with custom cloud role settings:");
+  console.log("Microsoft OpenTelemetry configured with custom cloud role settings:");
   console.log(
     `  Cloud Role Name: ${process.env.OTEL_SERVICE_NAMESPACE || "my-namespace"}.${process.env.OTEL_SERVICE_NAME || "my-service"}`,
   );
