+++ samples/src/liveMetrics.ts
@@ -5,21 +5,21 @@
  * @summary Demonstrates how to enable or disable Live Metrics for real-time monitoring.
  */
 
-import { useAzureMonitor } from "@azure/monitor-opentelemetry";
+import { useMicrosoftOpenTelemetry } from "@microsoft/opentelemetry";
 import "dotenv/config";
 
 async function main(): Promise<void> {
-  const options = {
-    azureMonitorExporterOptions: {
-      connectionString:
-        process.env.APPLICATIONINSIGHTS_CONNECTION_STRING || "<your connection string>",
+  useMicrosoftOpenTelemetry({
+    azureMonitor: {
+      azureMonitorExporterOptions: {
+        connectionString:
+          process.env.APPLICATIONINSIGHTS_CONNECTION_STRING || "<your connection string>",
+      },
+      enableLiveMetrics: true,
     },
-    enableLiveMetrics: true,
-  };
+  });
 
-  useAzureMonitor(options);
-
-  console.log("Azure Monitor configured with Live Metrics:");
+  console.log("Microsoft OpenTelemetry configured with Live Metrics:");
   console.log("  Live Metrics: ENABLED");
   console.log("  Check Azure Portal > Application Insights > Live Metrics Stream");
 }
