+++ samples/src/sampling.ts
@@ -5,22 +5,22 @@
  * @summary Demonstrates how to enable sampling to reduce data ingestion volume and control costs.
  */
 
-const { useAzureMonitor } = require("@azure/monitor-opentelemetry");
-require("dotenv/config");
+import { useMicrosoftOpenTelemetry } from "@microsoft/opentelemetry";
+import "dotenv/config";
 
-async function main() {
+async function main(): Promise<void> {
   // A sampling ratio of 0.1 means approximately 10% of traces are sent
-  const options = {
+  useMicrosoftOpenTelemetry({
     samplingRatio: 0.1,
-    azureMonitorExporterOptions: {
-      connectionString:
-        process.env.APPLICATIONINSIGHTS_CONNECTION_STRING || "<your connection string>",
+    azureMonitor: {
+      azureMonitorExporterOptions: {
+        connectionString:
+          process.env.APPLICATIONINSIGHTS_CONNECTION_STRING || "<your connection string>",
+      },
     },
-  };
+  });
 
-  useAzureMonitor(options);
-
-  console.log("Azure Monitor configured with sampling:");
+  console.log("Microsoft OpenTelemetry configured with sampling:");
   console.log("  Sampling Ratio: 10% (0.1)");
   console.log("  This reduces data ingestion volume and costs");
 }
